rag-application/requirements.txt
streamlit>=1.32.0
groq>=0.30.0
sentence-transformers>=2.5.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
numpy>=1.24.0
PyPDF2>=3.0.0
